Implement the sandbox directive for CSP https://bugs.webkit.org/show_bug.cgi?id=71604
Reviewed by Sam Weinig.

Source/WebCore:

At TPAC, Microsoft was pushing pretty hard to add the sandbox directive
to CSP.  There's a question about whether it's going to be in CSP 1.0
or CSP 1.1, but it seems to be clearly headed into the spec.

This patch implements the sandbox directive for CSP.  It's built on the
same machinery we use for the sandbox attribute for iframe.  Now that
I've done the implementation, I'm going to write up some concrete text
for the spec.

Tests: http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-subframe.html
       http/tests/security/contentSecurityPolicy/sandbox-allow-scripts.html
       http/tests/security/contentSecurityPolicy/sandbox-empty-subframe.html
       http/tests/security/contentSecurityPolicy/sandbox-empty.html

* page/ContentSecurityPolicy.cpp:
(WebCore::ContentSecurityPolicy::ContentSecurityPolicy):
(WebCore::ContentSecurityPolicy::applySandboxPolicy):
(WebCore::ContentSecurityPolicy::addDirective):
* page/ContentSecurityPolicy.h:

LayoutTests:

Test that the sandbox directive correctly wires up to the sandbox
flags.  Also, test that the policy inherits into subframes.  These are
essentially integration tests between the CSP and sandbox subsystems,
which is why they don't cover all the permutations of the sandbox
flags.  Those are covered in more detail via tests of the sandbox
attribute.

* http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-subframe-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-subframe.html: Added.
* http/tests/security/contentSecurityPolicy/sandbox-allow-scripts.html: Added.
* http/tests/security/contentSecurityPolicy/sandbox-empty-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/sandbox-empty-subframe-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/sandbox-empty-subframe.html: Added.
* http/tests/security/contentSecurityPolicy/sandbox-empty.html: Added.

Identifier: 87964@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99383 268f45cc-cd09-0410-ab3c-d52691b4dbfc