iframe sandbox treats vertical tab as a valid delimiter https://bugs.webkit.org/show_bug.cgi?id=71704
Reviewed by Eric Seidel.

Source/WebCore:

This patch adjusts our parser slightly to match the HTML5 spec.  The
only difference is in how we handle vertical tabs.  Previously, we
treated them as a delimiter, but we're not supposed to do that.

Test: fast/frames/sandboxed-iframe-parsing-space-characters.html

* page/SecurityOrigin.cpp:
(WebCore::SecurityOrigin::parseSandboxPolicy):

LayoutTests:

Test which space-like charaters are treating as delimiters.

* fast/frames/sandboxed-iframe-parsing-space-characters-expected.txt: Added.
* fast/frames/sandboxed-iframe-parsing-space-characters.html: Added.


Identifier: 88043@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99466 268f45cc-cd09-0410-ab3c-d52691b4dbfc