[Qt] Workaround some problems in QML API tests https://bugs.webkit.org/show_bug.cgi?id=71818
Reviewed by Tor Arne Vestbø.

This commit workaround the issue of QML's QtTest crashing when we run tests after
"windowShown". We delay to the next run of the mainloop. This workaround is not
pretty but shouldn't affect the semantics of the tests. And we benefit more from
having working tests right now.

This patch also improves Download to: wait for the load before emitting click, and
properly identifying the "succeeded" signal of the downloadItem object.

* UIProcess/API/qt/tests/qmltests/DesktopWebView/tst_download.qml:
* UIProcess/API/qt/tests/qmltests/DesktopWebView/tst_linkHovered.qml:
* UIProcess/API/qt/tests/qmltests/DesktopWebView/tst_navigationPolicyForUrl.qml:

Identifier: 88143@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99570 268f45cc-cd09-0410-ab3c-d52691b4dbfc