[Mac] Disable C++ exceptions from DumpRenderTree https://bugs.webkit.org/show_bug.cgi?id=71834
Patch by Benjamin Poulain <bpoulain@apple.com> on 2011-11-08
Reviewed by Darin Adler.

DumpRenderTree Mac does not handle C++ exceptions. When such exception
happens, it tends to cause the AutoReleasePool stack to be corrupted,
which leads to crashes long after the exception was raised.

This patch disabled C++ exceptions in order to catch the problems where
they happen when a C++ exception is raised.

* DumpRenderTree/mac/Configurations/DumpRenderTree.xcconfig:

Identifier: 88193@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99626 268f45cc-cd09-0410-ab3c-d52691b4dbfc