Implement dumpProgressFinishedCallback() for Mac layoutTestController https://bugs.webkit.org/show_bug.cgi?id=66772
Patch by Benjamin Poulain <bpoulain@apple.com> on 2011-11-09
Reviewed by Simon Fraser.

Tools:

When dumpProgressFinishedCallback is enabled, DumpRenderTree should output
"postProgressFinishedNotification" on FrameLoaderClient::postProgressFinishedNotification().

On Mac, the FrameLoaderClient post a notification. This patch add handling for this notification
in the FrameLoadDelegate and print the output in response to the notification.

* DumpRenderTree/mac/FrameLoadDelegate.mm:
(-[FrameLoadDelegate init]):
(-[FrameLoadDelegate dealloc]):
(-[FrameLoadDelegate webViewProgressFinishedNotification:]):

LayoutTests:

Unskip the skipped test.

* platform/mac/Skipped:

Identifier: 88327@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99762 268f45cc-cd09-0410-ab3c-d52691b4dbfc