ASSERT when outdenting styled line break. https://bugs.webkit.org/show_bug.cgi?id=72069 <rdar://problem/10429739>
Reviewed by Ryosuke Niwa.

Source/WebCore:

When outdenting a line break that has some inline style, we don't consider
the enclosing elements that provide the style.
This causes the splitting to occur in the wrong position,
triggering the assert.

Test: editing/execCommand/outdent-break-with-style.html

* editing/IndentOutdentCommand.cpp:
(WebCore::IndentOutdentCommand::outdentParagraph): The fix consists
in choosing the highest inline ancestor as splitting point.
* editing/htmlediting.cpp:
(WebCore::isInline): Added.
* editing/htmlediting.h: Added isInline.

LayoutTests:

* editing/execCommand/outdent-break-with-style-expected.txt: Added.
* editing/execCommand/outdent-break-with-style.html: Added.



Identifier: 88479@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99918 268f45cc-cd09-0410-ab3c-d52691b4dbfc