[chromium] Fix CCLayerTreeHost leak in webkit_unit_tests https://bugs.webkit.org/show_bug.cgi?id=72081
Patch by James Robinson <jamesr@chromium.org> on 2011-11-11
Reviewed by Kenneth Russell.

LayerChromium and CCLayerTreeHosts hold RefPtr<>s to each other, so if
you do not clear out the layerTreeHost pointer on the LayerChromium
the layer tree is leaked. The real code does this in WebViewImpl but
this test was not.

* tests/CCLayerTreeHostTest.cpp:
(WTF::CCLayerTreeHostTest::runTest):
(WTF::CCLayerTreeHostTest::doBeginTest):

Identifier: 88503@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99943 268f45cc-cd09-0410-ab3c-d52691b4dbfc