[chromium] Remove incorrect render surface layer list asserts from compositor https://bugs.webkit.org/show_bug.cgi?id=72744
Reviewed by James Robinson.

These asserts are incorrect, because it's valid for the default render
surface to be created but to have a root layer that doesn't draw. It
won't get added to the single render surface and it'll have an empty
layer list.

Test: compositor_unittests

* platform/graphics/chromium/LayerRendererChromium.cpp:
(WebCore::LayerRendererChromium::drawLayersOntoRenderSurfaces):
* platform/graphics/chromium/cc/CCLayerTreeHost.cpp:
(WebCore::CCLayerTreeHost::paintLayerContents):
(WebCore::CCLayerTreeHost::updateCompositorResources):


Identifier: 89436@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100947 268f45cc-cd09-0410-ab3c-d52691b4dbfc