[Qt] Re-generate QtWebKit API forwarding headers when API changes
The forwarding headers are generated by syncqt, and syncqt is run by qmake,
so we need sync.profile to live in the same directory as the project file
that will be touched when adding/removing API (api.pri in our case).

Since the WebKit2 APIs live in a separate static library, we have to
add the project file for WebKit2 as a dependency of the api.pri file
as well, so that any changes to the WebKit2 API will still trigger a
run of qmake (and then syncqt).

Lastly, we use the new QMAKE_SYNCQT_OUTDIR variable in Qt 5 to ensure
that the forwarding headers are still generated in the root build dir,
not in the Source dir along with the internal forwarding headers.

Reviewed by Andreas Kling.

Identifier: 89572@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101089 268f45cc-cd09-0410-ab3c-d52691b4dbfc